server/toota/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==23.12.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
cloudinary==1.42.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
coverage==7.3.3
cryptography==42.0.3
daphne==4.1.0
defusedxml==0.8.0rc2
dj-database-url==2.1.0
django-extensions==3.2.3
Django==5.0.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==2.2.0
django-phonenumber-field==7.2.0
django-templated-mail==1.1.1
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
exceptiongroup==1.2.0
frozenlist==1.4.1
git-filter-repo==2.38.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.26
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-django==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
##Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
zope.interface==6.2
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==23.12.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.42.2
colorama==0.4.6
constantly==23.10.4
coverage==7.3.3
cryptography==42.0.3
daphne==4.1.0
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.2
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-js-asset==2.2.0
django-phonenumber-field==7.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
exceptiongroup==1.2.0
frozenlist==1.4.1
git-filter-repo==2.38.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.26
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-django==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
##Twisted==23.10.0
##twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
zope.interface==6.2
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==23.12.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.42.2
colorama==0.4.6
constantly==23.10.4
coverage==7.3.3
cryptography==42.0.3
daphne==4.1.0
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.2
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-js-asset==2.2.0
django-phonenumber-field==7.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
exceptiongroup==1.2.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.38.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.26
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-django==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
zope.interface==6.2
